+++ include/m/vec/dvec3.h
@@ -428,7 +428,7 @@ static dvec3 dvec3_clamp_vec(dvec3 x, dvec3 min, dvec3 max) {
 
 
 /** dst = a * (1-t) + b * t */
-static dvec3 dvec3_mix(dvec3 a, dvec3 b, double t) {
+static dvec3 dvec3_mix(dvec3 a, dvec3 b, float t) {
     dvec3 res;
     dvecN_mix(res.v, a.v, b.v, t, 3);
     return res;
